src/main/java/lpoo2021/g64PvZ/model/Position.java
package lpoo2021.g64PvZ.model;

import java.util.Objects;

public class Position {
    private final int x;
    private final int y;

    public Position(int x, int y) {
        this.x = x;
        this.y = y;
    }

    public Position getLeft() {
        return new Position(x - 1, y);
    }

    public Position cursorGetLeft() { return new Position(x - 1, y); }

    public Position cursorGetRight() {
        return new Position(x + 1, y);
    }

    public Position cursorGetUp() {
        return new Position(x, y - 1);
    }

    public Position cursorGetDown() {
        return new Position(x, y + 1);
    }

    public int getX() {
        return x;
    }

    public int getY() {
        return y;
    }

    @Override
    public boolean equals(Object o) {
        if (this == o) return true;
        if (o == null || getClass() != o.getClass()) return false;
        Position position = (Position) o;
        return x == position.x && y == position.y;
    }

    @Override
    public int hashCode() {
        return Objects.hash(x, y);
    }
}